moving map out of app.py
