+++ src/callbacks/__init__.py
@@ -1,3 +1,4 @@
 # pull in components from files in the current directory to make imports cleaner
 
 from .bar_chart import update_bar_chart
+from .map import update_map
