+++ src/components/__init__.py
@@ -1,3 +1,4 @@
 # pull in components from files in the current directory to make imports cleaner
 
 from .bar_chart import bar_chart
+from .map import park_map, create_markers
